Add File::getChunkSize() and File::setChunkSize() no-op
stream_set_chunk_size() now uses setChunkSize()
Additional tests for stream_set_chunk_size()

+++ hphp/runtime/ext/stream/ext_stream.h
@@ -223,7 +223,7 @@ int64_t HHVM_FUNCTION(stream_set_read_buffer,
 
 Variant HHVM_FUNCTION(stream_set_chunk_size,
                       const Resource& stream,
-                      int chunk_size);
+                      int64_t chunk_size);
 
 bool HHVM_FUNCTION(stream_set_timeout,
                    const Resource& stream,
